requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
grequests==0.3.0
pymongo==3.10.1
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.13.0
regex==2019.4.14
requests==2.18.4
ruamel-yaml==0.15.35
secure-smtplib==0.1.1
Send2Trash==1.4.2
unicodecsv==0.14.1
urllib3==1.22
